# backend/public/swagger.json
{
	"components": {
...
		"schemas": {
			"Song": {
				"properties": {
					"song_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"song_user_id": {
						"type": "number",
						"format": "double"
					},
					"song_user_uuid": {
						"type": "string"
					},
					"file_path": {
						"type": "string"
					},
					"song_title": {
						"type": "string"
					},
					"artist": {
						"type": "string"
					},
					"album": {
						"type": "string"
					},
					"song_is_deleted": {
						"type": "number",
						"format": "double"
					},
					"song_created": {
						"type": "number",
						"format": "double"
					},
					"song_modified": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"song_id",
					"song_user_id",
					"song_user_uuid",
					"file_path",
					"song_title",
					"artist",
					"album",
					"song_is_deleted",
					"song_created",
					"song_modified"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Playlist": {
				"properties": {
					"playlist_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"playlist_user_id": {
						"type": "number",
						"format": "double"
					},
					"playlist_user_uuid": {
						"type": "string"
					},
					"playlist_title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"songs": {
						"items": {
							"$ref": "#/components/schemas/Song"
						},
						"type": "array",
						"nullable": true
					},
					"playlist_is_deleted": {
						"type": "number",
						"format": "double"
					},
					"playlist_created": {
						"type": "number",
						"format": "double"
					},
					"playlist_modified": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"playlist_id",
					"playlist_user_id",
					"playlist_user_uuid",
					"playlist_title",
					"description",
					"songs",
					"playlist_is_deleted",
					"playlist_created",
					"playlist_modified"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponsePlaylistGetAll": {
				"properties": {
					"is_success": {
						"type": "boolean"
					},
					"error_code": {
						"type": "number",
						"format": "double"
					},
					"error_msg": {
						"type": "string"
					},
					"playlists": {
						"items": {
							"$ref": "#/components/schemas/Playlist"
						},
						"type": "array"
					}
				},
				"required": [
					"is_success",
					"error_code",
					"error_msg",
					"playlists"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestPlaylistGetAll": {
				"properties": {
					"user_uuid": {
						"type": "string"
					},
					"session_hash": {
						"type": "string"
					},
					"request_hash": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"user_uuid",
					"session_hash",
					"request_hash",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponsePlaylistGetUser": {
				"properties": {
					"is_success": {
						"type": "boolean"
					},
					"error_code": {
						"type": "number",
						"format": "double"
					},
					"error_msg": {
						"type": "string"
					},
					"playlists": {
						"items": {
							"$ref": "#/components/schemas/Playlist"
						},
						"type": "array"
					}
				},
				"required": [
					"is_success",
					"error_code",
					"error_msg",
					"playlists"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestPlaylistGetUser": {
				"properties": {
					"user_uuid": {
						"type": "string"
					},
					"session_hash": {
						"type": "string"
					},
					"request_hash": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"user_uuid",
					"session_hash",
					"request_hash",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponsePlaylistCreate": {
				"properties": {
					"is_success": {
						"type": "boolean"
					},
					"error_code": {
						"type": "number",
						"format": "double"
					},
					"error_msg": {
						"type": "string"
					},
					"playlist_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"is_success",
					"error_code",
					"error_msg",
					"playlist_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestPlaylistCreate": {
				"properties": {
					"user_uuid": {
						"type": "string"
					},
					"session_hash": {
						"type": "string"
					},
					"request_hash": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"user_uuid",
					"session_hash",
					"request_hash",
					"timestamp",
					"title",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestPlaylistDelete": {
				"properties": {
					"user_uuid": {
						"type": "string"
					},
					"session_hash": {
						"type": "string"
					},
					"request_hash": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"playlist_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"user_uuid",
					"session_hash",
					"request_hash",
					"timestamp",
					"playlist_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseSongGetPlaylist": {
				"properties": {
					"is_success": {
						"type": "boolean"
					},
					"error_code": {
						"type": "number",
						"format": "double"
					},
					"error_msg": {
						"type": "string"
					},
					"songs": {
						"items": {
							"$ref": "#/components/schemas/Song"
						},
						"type": "array"
					}
				},
				"required": [
					"is_success",
					"error_code",
					"error_msg",
					"songs"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestSongGetPlaylist": {
				"properties": {
					"user_uuid": {
						"type": "string"
					},
					"session_hash": {
						"type": "string"
					},
					"request_hash": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"playlist_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"user_uuid",
					"session_hash",
					"request_hash",
					"timestamp",
					"playlist_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseSongGet": {
				"properties": {
					"is_success": {
						"type": "boolean"
					},
					"error_code": {
						"type": "number",
						"format": "double"
					},
					"error_msg": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"artist": {
						"type": "string"
					},
					"album": {
						"type": "string"
					},
					"file_b64": {
						"type": "string"
					}
				},
				"required": [
					"is_success",
					"error_code",
					"error_msg",
					"title",
					"artist",
					"album",
					"file_b64"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestSongGet": {
				"properties": {
					"user_uuid": {
						"type": "string"
					},
					"session_hash": {
						"type": "string"
					},
					"request_hash": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"song_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"user_uuid",
					"session_hash",
					"request_hash",
					"timestamp",
					"song_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseSongAdd": {
				"properties": {
					"is_success": {
						"type": "boolean"
					},
					"error_code": {
						"type": "number",
						"format": "double"
					},
					"error_msg": {
						"type": "string"
					},
					"song_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"is_success",
					"error_code",
					"error_msg",
					"song_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestSongAdd": {
				"properties": {
					"user_uuid": {
						"type": "string"
					},
					"session_hash": {
						"type": "string"
					},
					"request_hash": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"playlist_id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"artist": {
						"type": "string"
					},
					"album": {
						"type": "string"
					},
					"file_b64": {
						"type": "string"
					}
				},
				"required": [
					"user_uuid",
					"session_hash",
					"request_hash",
					"timestamp",
					"playlist_id",
					"title",
					"artist",
					"album",
					"file_b64"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseSongAddToPlaylist": {
				"properties": {
					"is_success": {
						"type": "boolean"
					},
					"error_code": {
						"type": "number",
						"format": "double"
					},
					"error_msg": {
						"type": "string"
					}
				},
				"required": [
					"is_success",
					"error_code",
					"error_msg"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestSongAddToPlaylist": {
				"properties": {
					"user_uuid": {
						"type": "string"
					},
					"session_hash": {
						"type": "string"
					},
					"request_hash": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"playlist_id": {
						"type": "number",
						"format": "double"
					},
					"song_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"user_uuid",
					"session_hash",
					"request_hash",
					"timestamp",
					"playlist_id",
					"song_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseSongDelete": {
				"properties": {
					"is_success": {
						"type": "boolean"
					},
					"error_code": {
						"type": "number",
						"format": "double"
					},
					"error_msg": {
						"type": "string"
					},
					"song_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"is_success",
					"error_code",
					"error_msg",
					"song_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestSongDelete": {
				"properties": {
					"user_uuid": {
						"type": "string"
					},
					"session_hash": {
						"type": "string"
					},
					"request_hash": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"song_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"user_uuid",
					"session_hash",
					"request_hash",
					"timestamp",
					"song_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ts-task-backend",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "pukanjourpour"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/playlists/all": {
			"post": {
				"operationId": "GetAllPlaylists",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponsePlaylistGetAll"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestPlaylistGetAll"
							}
						}
					}
				}
			}
		},
		"/playlists/user": {
			"post": {
				"operationId": "GetUserPlaylists",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponsePlaylistGetUser"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestPlaylistGetUser"
							}
						}
					}
				}
			}
		},
		"/playlists/create": {
			"post": {
				"operationId": "CreatePlaylist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponsePlaylistCreate"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestPlaylistCreate"
							}
						}
					}
				}
			}
		},
		"/playlists/delete": {
			"post": {
				"operationId": "DeletePlaylist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponsePlaylistCreate"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestPlaylistDelete"
							}
						}
					}
				}
			}
		},
		"/songs/get": {
			"post": {
				"operationId": "GetPlaylistSongs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseSongGetPlaylist"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestSongGetPlaylist"
							}
						}
					}
				}
			}
		},
		"/songs/get-one": {
			"post": {
				"operationId": "GetSong",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseSongGet"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestSongGet"
							}
						}
					}
				}
			}
		},
		"/songs/add": {
			"post": {
				"operationId": "AddSong",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseSongAdd"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestSongAdd"
							}
						}
					}
				}
			}
		},
		"/songs/add-to-playlist": {
			"post": {
				"operationId": "AddSongToPlaylist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseSongAddToPlaylist"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestSongAddToPlaylist"
							}
						}
					}
				}
			}
		},
		"/songs/delete": {
			"post": {
				"operationId": "DeleteSong",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseSongDelete"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestSongDelete"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}